create_executor* functions return global executors
This is necessary because we implicitly assume that a program only has one executor of a specific type (up to changes in construtor parameters). Thus, without this we create different executors each time the create_* functions are called. And as a consequence extra copies are used instead of moves:
```
auto exec1 = create_default_executor();
auto exec2 = create_default_executor();
array<...> a{exec1, ...};
array<...> b{std::move(a)}; // this copies!
```
Alternative: move the responsibility to downstream projects.